# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatsapp Api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://official.whatsapp.loc/engine"
        }
    ],
    "paths": {
...
        "/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "fetch all chats",
                "description": "fetch user all chats",
                "operationId": "chats",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "query",
                        "description": "Chat Id to get specific chat",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Page Size",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns array of chats."
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/deleteChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "delete chat",
                "description": "delete specific chat",
                "operationId": "deleteChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be deleted",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/readChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "read chat",
                "description": "read specific chat",
                "operationId": "readChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be read",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/unreadChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "unread chat",
                "description": "read specific chat",
                "operationId": "unreadChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be unread",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/archiveChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "archive chat",
                "description": "archive specific chat",
                "operationId": "archiveChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be archived",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/unarchiveChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "unarchive chat",
                "description": "unarchive specific chat",
                "operationId": "unarchiveChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be unarchived",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/pinChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "pin chat",
                "description": "pin specific chat",
                "operationId": "pinChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be pinned",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/unpinChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "unpin chat",
                "description": "unpin specific chat",
                "operationId": "unpinChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be unpinned",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/muteChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "mute chat",
                "description": "mute specific chat",
                "operationId": "muteChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be muted",
                        "required": true
                    },
                    {
                        "name": "duration",
                        "in": "query",
                        "description": "duration of seconds that chat would be muted",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
        "/chats/unmuteChat": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "unmute chat",
                "description": "unmute specific chat",
                "operationId": "unmuteChat",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be unmuted",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
        "/chats/recording": {
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "recording",
                "description": "set recording status to specific chat",
                "operationId": "recording",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "chat phone to be notified about your status",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer_token": [],
                        "channel_id": [],
                        "channel_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "channel_id": {
                "type": "apiKey",
                "description": "Enter Your Channel ID",
                "name": "CHANNELID",
                "in": "header"
            },
            "channel_token": {
                "type": "apiKey",
                "description": "Enter Your Channel Token",
                "name": "CHANNELTOKEN",
                "in": "header"
            }
        }
    }
}